fix: change severity of missing file-path message to debug
When the user has indicated that a missing file is not an error
(because all settings have defaults), we emit a "file-path was not
found" message that attaches to the workflow by default.

Change the severity level of the message from "notice" to "debug", to
clear a non-error message from the workflow unless the user has asked
for debug verbosity.